requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
openai==1.10.0
requests==2.31.0
python-dateutil==2.8.2
pydantic==2.5.3
httpx==0.26.0
